Fix crud for todos, Add MUI & AppBar

+++ app/todos/mutations/deleteTodo.ts
@@ -6,9 +6,13 @@ const DeleteTodo = z.object({
   id: z.number(),
 })
 
-export default resolver.pipe(resolver.zod(DeleteTodo), resolver.authorize(), async ({ id }) => {
-  // TODO: in multi-tenant app, you must add validation to ensure correct tenant
-  const todo = await db.todo.deleteMany({ where: { id } })
+export default resolver.pipe(
+  resolver.zod(DeleteTodo),
+  resolver.authorize(),
+  async ({ id }, ctx) => {
+    // TODO: in multi-tenant app, you must add validation to ensure correct tenant
+    const todo = await db.todo.deleteMany({ where: { id, authorId: ctx.session.userId } })
 
-  return todo
-})
+    return todo
+  }
+)
